+++ smol/utils/cluster/container.pyx
@@ -85,8 +85,8 @@ cdef class OrbitContainer:
     cdef OrbitC create_struct(
             int orbit_id,
             int bit_id,
-            double[:, ::1] correlation_tensors,
-            long[::1] tensor_indices,
+            np.float64_t[:, ::1] correlation_tensors,
+            np.int32_t[::1] tensor_indices,
     ):
         """Set the fields of a OrbitC struct from memoryviews."""
         cdef OrbitC orbit
@@ -154,7 +154,7 @@ cdef class FloatArray2DContainer:
 
 
     @staticmethod
-    cdef FloatArray2D create_struct(double[:, ::1] array):
+    cdef FloatArray2D create_struct(np.float64_t[:, ::1] array):
         """Set the fields of a _FloatArray2D struct from memoryview."""
         cdef FloatArray2D array_struct
         array_struct.size_r = array.shape[0]
@@ -212,7 +212,7 @@ cdef class FloatArray1DContainer:
         self._arrays = arrays
 
     @staticmethod
-    cdef FloatArray1D create_struct(double[::1] array):
+    cdef FloatArray1D create_struct(np.float64_t[::1] array):
         """Set the fields of a FloatArray1D struct from memoryview."""
         cdef FloatArray1D array_struct
         array_struct.size = array.size
@@ -269,7 +269,7 @@ cdef class IntArray1DContainer:
         self._arrays = arrays
 
     @staticmethod
-    cdef IntArray1D create_struct(long[::1] array):
+    cdef IntArray1D create_struct(np.int32_t[::1] array):
         """Set the fields of a _FloatArray1D struct from memoryview."""
         cdef IntArray1D array_struct
         array_struct.size = array.size
@@ -326,7 +326,7 @@ cdef class IntArray2DContainer:
         self._arrays = arrays
 
     @staticmethod
-    cdef IntArray2D create_struct(long[:, ::1] array):
+    cdef IntArray2D create_struct(np.int32_t[:, ::1] array):
         """Set the fields of a _IntArray2D struct from memoryview."""
         cdef IntArray2D array_struct
         array_struct.size_r = array.shape[0]
